pyproject.toml
[tool.poetry]
name = "iching"
version = "0.1.0"
description = "AI Fortune Teller"
authors = ["jiajiabinx"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
wikipedia-api = "^0.7.1"
numpy = "1.26.3"
beautifulsoup4 = "^4.12.3"
sentence-transformers = "^3.0.1"
ipywidgets = "^8.1.5"
faiss-cpu = "^1.8.0.post1"
fastapi = "^0.114.1"
pydantic = "^2.9.1"
v = "^1"
uvicorn = "^0.30.6"
python-multipart = "^0.0.9"
jinja2 = "^3.1.4"
glob2 = "^0.7"
python-dotenv = "^1.0.1"
anthropic = "^0.34.2"
sqlalchemy = "^2.0.36"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
